backend/requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
jinja2==3.1.4
python-dotenv==0.21.1
websocket-client==1.8.0
loguru==0.7.2
httpx==0.27.2
pydantic==1.10.13
google-cloud-dialogflow-cx==1.34.1
google-auth==2.35.0
google-cloud-speech==2.26.0
